% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hierarchy.R
\name{add_level_indicator}
\alias{add_level_indicator}
\title{Attach hidden levels indicators to the main application data.}
\usage{
add_level_indicator(data, subfoldable_levels, foldable_label, levels_meta)
}
\arguments{
\item{data}{Joined Data and Indicator data, storing the scores.}

\item{subfoldable_levels}{Levels with the hierarchy lower than foldable.}

\item{levels_meta}{Levels meta information object.}
}
\description{
Attach hidden levels indicators to the main application data.
}
