% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generate.R
\name{aggregate_scores}
\alias{aggregate_scores}
\title{Aggregate level-based scores}
\usage{
aggregate_scores(aggregation_data, levels_meta, method = "DEFAULT", ...)
}
\arguments{
\item{aggregation_data}{Data storing scores, and weights for each hierarchy level.}

\item{levels_meta}{Hierarchy levels meta information object.}

\item{score_column}{Name of the column that stores scores to be aggregated.}
}
\description{
The function takes original scores data and levels meta information, and performs 
aggregation for all weighted levels that were defined in hierarchy Sheet.
}
