% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{all_equal}
\alias{all_equal}
\title{Check whether vector stores constant values}
\usage{
all_equal(vector, omit)
}
\arguments{
\item{vector}{Checked vector.}

\item{omit}{Exception values to be skipped in check (like 0, "-1")}
}
\description{
Check whether vector stores constant values
}
