% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatif_logic.R
\name{attach_ui_element}
\alias{attach_ui_element}
\alias{attach_ui_elements}
\title{Generate UI Element and attach it's code to a column}
\usage{
attach_ui_element(
  ui_element,
  selector_1,
  selector_2,
  line_id,
  numans,
  foldable_id,
  list_data
)

attach_ui_elements(
  ui_element,
  selector_1,
  selector_2,
  line_id,
  numans,
  foldable_id,
  list_data
)
}
\arguments{
\item{ui_element}{UI Element name.}

\item{selector_1}{ID of chosen selector 1.}

\item{selector_2}{ID of chosen selector 2.}

\item{line_id}{Line ID of corresponding UI Element.}

\item{numans}{Initial UI Element value (numans).}

\item{foldable_id}{ID of foldable level in which UI Element should be generated.}

\item{list_data}{List sheet data content.}
}
\description{
\code{attach_ui_element} Generates UI Element for a single data row.
\code{attach_ui_element} Vectorizes version of \code{attach_ui_element}.
}
