% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatif_logic.R
\name{calc_score}
\alias{calc_score}
\title{Extract scores based on evaluated conditions}
\usage{
calc_score(data, default = 0, score_col = "Score")
}
\arguments{
\item{data}{Data containing evaluated conditions ("Cond*_evaluated", and related scores).}

\item{default}{Default score when all the evaluated conditions are FALSE.}

\item{score_col}{Name of the column in which to store extracted scores.}
}
\description{
Extract scores based on evaluated conditions
}
