% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{case_in_set}
\alias{case_in_set}
\title{Case insensitive version of \link[assertr]{in_set}}
\usage{
case_in_set(...)
}
\arguments{
\item{...}{Objects that make up the set}
}
\description{
Case insensitive version of \link[assertr]{in_set}
}
