% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{isNumeric}
\alias{isNumeric}
\alias{Isnumeric}
\title{Alternative function names for \code{is.numeric}}
\usage{
isNumeric()

Isnumeric()
}
\arguments{
\item{x}{object to be coerced to numeric}
}
\description{
Alternative function names for \code{is.numeric}
}
\details{
used for correct evaluation of condition rules and ui elements conditions
}
