% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatif_logic.R
\name{compute_score}
\alias{compute_score}
\title{Apply computing the scores logic}
\usage{
compute_score(
  country_indicator_data,
  numals_env,
  levels_meta,
  score_col = "Score",
  update = FALSE
)
}
\arguments{
\item{country_indicator_data}{Joined Data and Indicator sheets data.}

\item{numals_env}{Environment storing line ID related numans.}

\item{levels_meta}{Hierarchy levels meta information object.}

\item{score_col}{Name of the column storing final score.}

\item{update}{If TRUE, additionally saves the computed scores to 'EditedValue' column.}
}
\description{
The function is responsible to perform all the required steps to extract proper condition scores.
The steps are:
1. Grouping the data based on selector1 and selector2.
2. Evaluation of conditions based on provided environment.
3. Extracting the score attached to the first met condition.
4. Assigning selected score to a target column.
}
