% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{conds_in_vars_set}
\alias{conds_in_vars_set}
\title{Check whether only defined set of variables is included in conditions set}
\usage{
conds_in_vars_set(conds_vec, set_regexp = "numans|num[:alnum:]*")
}
\arguments{
\item{conds_vec}{Vector of conditions.}

\item{set_regexp}{Regular expression defining variables set.}
}
\description{
By default only \code{numans} and \code{num_<lineID>} are allowed
}
