% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatif_logic.R
\name{conds_to_vec}
\alias{conds_to_vec}
\title{Convert conditions column to a single string prepared for evaluation.}
\usage{
conds_to_vec(conditions_vector)
}
\arguments{
\item{conditions_vector}{Vector with conditions.}
}
\description{
Convert conditions column to a single string prepared for evaluation.
}
