% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{const_per_group}
\alias{const_per_group}
\title{Helper function that returns provided value}
\usage{
const_per_group(value)
}
\arguments{
\item{value}{Value to be returned.}
}
\description{
Helper function that returns provided value
}
\details{
Used for making results of constants check more readable
}
