% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{construct_numans_in_list}
\alias{construct_numans_in_list}
\title{Helper predicate constructor for validation of numans offered by UI Element}
\usage{
construct_numans_in_list(list_data, ui_col = "UI Element")
}
\arguments{
\item{list_data}{List data sheet content.}

\item{ui_col}{Name of the column where UI Element names are stored in the target data.}
}
\description{
Helper predicate constructor for validation of numans offered by UI Element
}
