% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generate.R
\name{convert_to_character}
\alias{convert_to_character}
\title{Convert selected columns to character}
\usage{
convert_to_character(data, columns = colnames(data))
}
\arguments{
\item{data}{Table object}

\item{columns}{Columns to convert}
}
\description{
Convert selected columns to character
}
