% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generate.R
\name{convert_to_upper}
\alias{convert_to_upper}
\title{Convert selected columns to upper values}
\usage{
convert_to_upper(data, columns = colnames(data)[-1])
}
\arguments{
\item{data}{Table object}

\item{columns}{Columns to convert}
}
\description{
Convert selected columns to upper values
}
