% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{correct_expr}
\alias{correct_expr}
\title{Check whether expression included in string is a proper R expression}
\usage{
correct_expr(text_expr)
}
\arguments{
\item{text_expr}{Text expression to be checks.}
}
\description{
Check whether expression included in string is a proper R expression
}
