% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_generate.R
\name{create_app_structure}
\alias{create_app_structure}
\title{Create Policy Simulator App structure}
\usage{
create_app_structure(template, target_app_path, data)
}
\arguments{
\item{template}{List of tables included in the original template Sheets.}

\item{data}{Preprocessed data based on \code{template} that should be directly used in the app.}

\item{target-app_path}{Target application path.}
}
\description{
The function created the target application code in desired directory.
}
