% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatif_logic.R
\name{eval_conds_vec}
\alias{eval_conds_vec}
\title{Evaluate vector of conditions stored in a string.}
\usage{
eval_conds_vec(conditions_vector, envir)
}
\arguments{
\item{conditions_vector}{Vector with conditions.}

\item{envir}{Environment of variables values that should be used in evaluation.}
}
\description{
Evaluate vector of conditions stored in a string.
}
