% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generate.R
\name{export_template_data}
\alias{export_template_data}
\title{Source template Excel file sheets and export to separate RDS files}
\usage{
export_template_data(template, app_path)
}
\arguments{
\item{template}{Template resulting data to be exported}

\item{app_path}{Path to target application directory}
}
\description{
Source template Excel file sheets and export to separate RDS files
}
