% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{extract_numlines}
\alias{extract_numlines}
\title{Extract all 'num<line_id>' strings from condition text}
\usage{
extract_numlines(condition)
}
\arguments{
\item{condition}{Any string defining scoring condition, like 'num123 < 2'}
}
\description{
Extract all 'num<line_id>' strings from condition text
}
