% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatif_logic.R
\name{first_true}
\alias{first_true}
\title{Extract which value from a logical vector is the first TRUE object}
\usage{
first_true(logical_vector)
}
\arguments{
\item{logical_vector}{Logical vector to be checked.}
}
\description{
Extract which value from a logical vector is the first TRUE object
}
