% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{fun_vectorize}
\alias{fun_vectorize}
\title{Generate vectorized function evaluation code}
\usage{
fun_vectorize(arg, func, holder = "c")
}
\arguments{
\item{arg}{Vector of arguments to be passed of the function}

\item{holder}{What type of output object should be returned (e.g. 'c' or 'list').}

\item{fun}{Name of the function to be evaluated.}
}
\description{
Generate vectorized function evaluation code
}
