% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{generate_simulator}
\alias{generate_simulator}
\title{Validate template and generate policy simulator app}
\usage{
generate_simulator(
  excel_template_path,
  target_app_path = getwd(),
  validation = TRUE
)
}
\arguments{
\item{excel_template_path}{Path to excel template file.}

\item{target_app_path}{Target path where policy simulator should be generated. 
Current working directory by default.}
}
\description{
Provide excel template file target application paths in order to generate policy simulator.
The template is automatically validated (see Details section to learn more).
}
\details{
Everytime you try to generate policy simulator, provided template is validated in two steps.

The first one, preprocessing validation, takes raw template data and validates it with extensive set of rules
(see \code{vignette("template", package = "PolicySimulator")} for more details).

Violation of any of rules, stops further policy generation steps and detailed report is 
generated in the target application path.

The second one, postprocessing validation, is run after the target application data is prepared.
This step checks whether generated scores matched the one provided in the template. 
Such check is only informative, it doesn't stop generation of policy simulator.
}
