% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{get_cond_vars}
\alias{get_cond_vars}
\title{Extract all variables used in expression}
\usage{
get_cond_vars(condition)
}
\arguments{
\item{condition}{Text vector of R expressions from which variables should be extracted.}
}
\description{
Extract all variables used in expression
}
