% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatif_logic.R
\name{get_level_col}
\alias{get_level_col}
\title{Get name of corresponding meta level attached to ID.}
\usage{
get_level_col(level, prefix = "L")
}
\arguments{
\item{level}{Level ID Name.}

\item{prefix}{Desired level prefix. "L" default.}
}
\description{
Get name of corresponding meta level attached to ID.
}
\examples{
get_level_col("CH1", "W")
# WH1

get_level_col("CH1")
# LH1

}
