% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{get_template}
\alias{get_template}
\title{Get Policy Simulator Template skeleton}
\usage{
get_template(template_dir = getwd())
}
\arguments{
\item{template_dir}{Path where the template should be saved.}
}
\description{
Get Policy Simulator Template skeleton
}
