% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{get_violated_data}
\alias{get_violated_data}
\title{Get violated data rows}
\usage{
get_violated_data(data, validator, name = NULL, save = TRUE, path = getwd())
}
\arguments{
\item{data}{Validated dataset.}

\item{validator}{Validator object in which validation results of \code{data} are stored.}

\item{name}{Validated data ID. Should be consistent with the one provided in \link[data.validator]{add_results}.}

\item{save}{Should the data be saved to csv file?}

\item{path}{Where the data should be saved?}
}
\description{
Get violated data rows
}
