% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hierarchy.R
\name{get_evaluate_level}
\alias{get_evaluate_level}
\alias{get_foldable_level}
\alias{get_hierarchy_levels}
\alias{get_hidden_levels}
\alias{get_weighted_levels}
\alias{get_selector_level}
\alias{get_levels_meta}
\title{Set of functions to extract levels set of specific types}
\usage{
get_evaluate_level(data)

get_foldable_level(data)

get_hierarchy_levels(data)

get_hidden_levels(data)

get_weighted_levels(data)

get_selector_level(data, selector_name)

get_levels_meta(data)
}
\arguments{
\item{data}{Hierarchy sheet dataset.}

\item{selector_name}{Name of the selector ("Selector1", "Selector2" or "CompareSelector")}
}
\description{
\code{get_levels_meta} Extracts a list of required levels meta in the package.
}
