% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generate.R
\name{import_simulator_data}
\alias{import_simulator_data}
\title{Import RDS Simulator base files}
\usage{
import_simulator_data(path = "data")
}
\arguments{
\item{path}{Path to simulator rds files}
}
\description{
Import RDS Simulator base files
}
