% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generate.R
\name{import_template_data}
\alias{import_template_data}
\title{Import template Excel file sheets as list}
\usage{
import_template_data(excel_template_path)
}
\arguments{
\item{excel_template_path}{Path to template excel file.}
}
\description{
Import template Excel file sheets as list
}
