% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is_equal}
\alias{is_equal}
\title{Are the two columns of the data equal}
\usage{
is_equal(data)
}
\arguments{
\item{data}{Data consisting of two columns.}
}
\description{
Are the two columns of the data equal
}
