% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatif_logic.R
\name{line_numans_to_env}
\alias{line_numans_to_env}
\title{Convert line ID numans to an environment.}
\usage{
line_numans_to_env(numans, line_id)
}
\arguments{
\item{numans}{Vector of line ID numans values.}

\item{line_id}{Vector of Line IDs.}
}
\description{
Convert line ID numans to an environment.
}
