% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hierarchy.R
\name{make_hierarchy_structure}
\alias{make_hierarchy_structure}
\title{Add hidden levels indicators to each foldable level and attach placeholders for future edited scored and values}
\usage{
make_hierarchy_structure(data, levels_meta)
}
\arguments{
\item{data}{Joined Data and Indicator data, storing the scores.}

\item{levels_meta}{Levels meta information object.}
}
\description{
Add hidden levels indicators to each foldable level and attach placeholders for future edited scored and values
}
