% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatif_logic.R
\name{match_cols}
\alias{match_cols}
\title{Select columns that match reactive expression}
\usage{
match_cols(data, reg_exp)
}
\arguments{
\item{data}{Data sect from which to extract matching column names.}

\item{reg_exp}{Regular expression from which to extract columns.}
}
\description{
Select columns that match reactive expression
}
