% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{num_in_lineid}
\alias{num_in_lineid}
\title{Check whether all conditions included in 'Code*' columns, fits in possible set of values defined in LineID column}
\usage{
num_in_lineid(data)
}
\arguments{
\item{data}{containing conditions (Cond* columns) and line ids (LineID column).}
}
\description{
Check whether all conditions included in 'Code*' columns, fits in possible set of values defined in LineID column
}
