% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatif_logic.R
\name{get_numals_selector_2}
\alias{get_numals_selector_2}
\alias{get_numals_vals}
\title{Generate line ID <-> numans environment for selector data.}
\usage{
get_numals_selector_2(per_selector_1_data, levels_meta)

get_numals_vals(country_indicator_data, levels_meta)
}
\arguments{
\item{per_selector_1_data}{Scores data for all sectors (selector1 values).}

\item{levels_meta}{Hierarchy levels meta information object.}
}
\description{
\code{get_numals_selector_2} creates environment for a single sector.
\code{get_numals_vals} iterates over all the selector_1 values and applies get_numals_selector_2.
}
