% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatif_logic.R
\name{numans_to_numline}
\alias{numans_to_numline}
\title{Convert 'numans' string in condition to a corresponding \code{num_{line_id}} one.}
\usage{
numans_to_numline(conditions_vector, line_numbers)
}
\arguments{
\item{conditions_vector}{Vector with conditions.}

\item{line_numbers}{Corresponding line ids.}
}
\description{
Convert 'numans' string in condition to a corresponding \code{num_{line_id}} one.
}
