% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{numline_in_lineid}
\alias{numline_in_lineid}
\title{Check whether all numlines exist in line ids set.}
\usage{
numline_in_lineid(num_lines, line_ids)
}
\arguments{
\item{num_lines}{Vector of \code{num_<lineid>} like values.}

\item{line_ids}{Desired set of possible numlines.}
}
\description{
Check whether all numlines exist in line ids set.
}
