% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{paste_cols}
\alias{paste_cols}
\title{Paste all columns of the data (with ' ' separator)}
\usage{
paste_cols(data)
}
\arguments{
\item{data}{Data which columns should be pasted}
}
\description{
Paste all columns of the data (with ' ' separator)
}
