% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hierarchy.R
\name{restructure_hierarchy}
\alias{restructure_hierarchy}
\title{Add hidden levels indicators to each foldable level}
\usage{
restructure_hierarchy(data, levels_meta)
}
\arguments{
\item{data}{Joined Data and Indicator data, storing the scores.}

\item{levels_meta}{Levels meta information object.}
}
\description{
Add hidden levels indicators to each foldable level
}
