% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is_radio}
\alias{is_radio}
\alias{is_list}
\alias{is_text}
\title{Check type of UI Element}
\usage{
is_radio(ui_element)

is_list(ui_element)

is_text(ui_element)
}
\arguments{
\item{ui_element}{UI element name}
}
\description{
Set of functions responsible for checking UI Element type.
Currently supported types:
\itemize{
  \item{radio buttons}{UI Element which name starts with "Radio"}
  \item{dropdown}{UI Element which name starts with "list"}
  \item{text input}{UI Element which name starts with "Text"}
}
}
