% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation_error_callback}
\alias{validation_error_callback}
\title{Define follow up steps on violated data}
\usage{
validation_error_callback(
  validator,
  target_app_path,
  template_file,
  callback = stop,
  report_file = "validation_report.html"
)
}
\arguments{
\item{validator}{Validator object.}

\item{target_app_path}{Target path when report should be saved.}

\item{template_file}{Validation template path.}

\item{callback}{Function that should be applied when violation exists. By default \code{stop}.}

\item{report_file}{Report file name.}
}
\description{
Define follow up steps on violated data
}
\details{
The function checks for existence of validation violations and performs 
the following (optional based on specified options) steps:
\itemize{
  \item{}{Saving the validation report when errors or warnings exist.}
  \item{}{Opens the report in the browser.}
  \item{}{Stops execution or returns warning about violations.}
}
}
