% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_hierarchy}
\alias{validate_hierarchy}
\alias{validate_list}
\alias{validate_data}
\alias{validate_indicator}
\alias{validate_answers}
\alias{validate_scores}
\title{Validate Template data}
\usage{
validate_hierarchy(hierarchy_data, validator, path)

validate_list(list_data, validator, path)

validate_data(country_data, validator, path)

validate_indicator(indicator_data, list_data, validator, levels_meta, path)

validate_answers(
  country_data,
  indicator_data,
  list_data,
  levels_meta,
  validator,
  path
)

validate_scores(score_data, levels_meta, validator, path)
}
\arguments{
\item{hierarchy_data}{Hierarchy sheet data content.}

\item{validator}{Validator object to store results in. See \link[data.validator]{create_valdiator} for more.}

\item{path}{Target path where violated data should be saved.}

\item{list_data}{List sheet data content.}

\item{country_data}{Data sheet data content.}

\item{indicator_data}{Indicator sheet data content.}

\item{levels_meta}{Metadata of hierarchy levels.}

\item{score_data}{Final application data storing original and computed scores.}
}
\description{
The below set of functions aims to check validity of the template data.
}
\details{
Each template Sheet, is validated by corresponding \code{validate_{sheet_name}} function.

Function \code{validate_answers} joins 'Data' and 'Indicator' sheets and validates the whole, joined dataset.

Function \code{validate_scores} is used during postprocessing validation check. 
It checks whether computed scores matches the one originally provided in the template.
}
