% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\docType{data}
\name{validation_report_template}
\alias{validation_report_template}
\title{Validation report template}
\format{
An object of class \code{character} of length 1.
}
\usage{
validation_report_template
}
\description{
Validation report template
}
\details{
Stored as a text due to lack of read/write permissions.
}
\keyword{datasets}
