% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{vals_in_set}
\alias{vals_in_set}
\title{Check whether all values are included in set}
\usage{
vals_in_set(vals, set, na_true = TRUE)
}
\arguments{
\item{vals}{Values which existence is checked in a set.}

\item{set}{Values set in which we check existence of values.}

\item{na_true}{If TRUE, then existence of NA in \code{vals} returns TRUE.}
}
\description{
Check whether all values are included in set
}
